vosemovies-backend/api/requirements.txt
# FastAPI Backend Requirements

# ========== Web Framework ==========
fastapi==0.115.5              # Modern async web framework
uvicorn[standard]==0.34.0     # ASGI server
pydantic==2.10.4              # Data validation
pydantic-settings==2.6.1      # Settings management

# ========== Database ==========
sqlalchemy==2.0.36            # ORM
psycopg2-binary==2.9.10       # PostgreSQL driver
alembic==1.14.0               # Database migrations

# ========== Utilities ==========
python-dotenv==1.0.1          # Environment variables
python-dateutil==2.9.0        # Date utilities
requests==2.31.0              # HTTP client for TMDB API

# ========== CORS & Security ==========
python-multipart==0.0.20      # Form data parsing
slowapi==0.1.9                # Rate limiting middleware

# ========== For scrapers integration ==========
beautifulsoup4==4.14.2        # HTML parsing (from scrapers)
selenium==4.38.0              # Browser automation (from scrapers)
